refactor: move configcenter -> pkg/ & delete igt/
