+++ pkg/config/config_load.go
@@ -36,8 +36,8 @@ import (
 )
 
 import (
-	"github.com/apache/dubbo-go-pixiu/configcenter"
 	"github.com/apache/dubbo-go-pixiu/pkg/common/constant"
+	"github.com/apache/dubbo-go-pixiu/pkg/configcenter"
 	"github.com/apache/dubbo-go-pixiu/pkg/logger"
 	"github.com/apache/dubbo-go-pixiu/pkg/model"
 )
